add signin

+++ backend/src/models/user.model.ts
@@ -7,47 +7,36 @@ import path from "path";
 const filepath = path.join(__dirname, "../../");
 const userDataFilePath = path.join(filepath, "data/users.json");
 
-function readUsers() {
+function getUsers() {
   const data = fs.readFileSync(userDataFilePath, "utf-8");
   return JSON.parse(data);
 }
 
 function createUser(user: Omit<User, "id">) {
-  const users = readUsers();
+  const users = getUsers();
+  const existingUser = users.find((u: User) => u.email === user.email);
+  if (existingUser) {
+    return null;
+  }
   const hashedPassword = bcrypt.hashSync(user.password, 10);
-  const newUser = { ...user, password: hashedPassword, id: uuidv4() };
+  const newUser = { ...user, password: hashedPassword, id: uuidv4(), role: "member" };
   users.push(newUser);
   fs.writeFileSync(userDataFilePath, JSON.stringify(users, null, 2));
   return newUser;
 }
 
 function getUserByEmail(email: string) {
-  const users = readUsers();
+  const users = getUsers();
   return users.find((user: User) => user.email === email);
 }
 
-function signinUser(email: string, password: string) {
-  const user = getUserByEmail(email);
-  if (!user) {
-    return null;
-  }
-  const isPasswordValid = bcrypt.compareSync(password, user.password);
-  if (!isPasswordValid) {
-    return null;
-  }
-  return user;
-}
-
-function signOutUser() {
-  return null;
-}
-
-function checkAuth() {
-  return null;
+function getUserById(id: string) {
+  const users = getUsers();
+  return users.find((user: User) => user.id === id);
 }
 
 function updateUser(id: string, user: User) {
-  const users = readUsers();
+  const users = getUsers();
   const userIndex = users.findIndex((user: User) => user.id === id);
   if (userIndex === -1) {
     return null;
@@ -59,7 +48,7 @@ function updateUser(id: string, user: User) {
 }
 
 function deleteUser(id: string) {
-  const users = readUsers();
+  const users = getUsers();
   const userIndex = users.findIndex((user: User) => user.id === id);
   if (userIndex === -1) {
     return null;
@@ -70,12 +59,10 @@ function deleteUser(id: string) {
 }
 
 export const userModel = {
-  readUsers,
+  getUsers,
   createUser,
   getUserByEmail,
-  signinUser,
-  signOutUser,
-  checkAuth,
+  getUserById,
   updateUser,
   deleteUser,
 };
